+++ lineprofiler/__init__.py
@@ -20,4 +20,4 @@ __all__ = [
     "stop_profiling",
 ]
 
-__version__ = "0.8.3"
+__version__ = "0.8.4"